Revert "attempt at using proper endianness"
This reverts commit 145aaca763fcfc2ac6a84a4a4c82c3db10410501.

+++ base-qt/BaseSocketInterface.cpp
@@ -55,7 +55,7 @@ void BaseSocketInterface::_sendPendingBuffers( void )
 //_______________________________________________________
 void BaseSocketInterface::_sendBuffer( qint32 type, const QByteArray& buffer )
 {
-    quint64 bufferSize( qToBigEndian( quint64(buffer.size()) ) );
+    quint64 bufferSize( qToBigEndian( buffer.size() ) );
     Debug::Throw(0) << "BaseSocketInterface::_sendBuffer - type: " << type << " size: " << bufferSize << endl;
     socket_->write( reinterpret_cast<const char*>( &type ), sizeof( qint32 ) );
     socket_->write( reinterpret_cast<const char*>( &bufferSize ), sizeof( quint64 ) );
